requirements.txt
torch<3
numpy<3
hydra-core
tqdm
transformers
datasets
tiktoken